Add files via upload
Here is the updated dislocation density based model (considering edge dislocations) for deformation in pure fcc metals (more specifically medium to high SFE fcc metals). The parameters used here are specifically for pure copper. The model has the temperature and strain rate dependency in it. The constitutive model is good enough to predict large strain deformation such as cold rolling. More details about the code will be given in the README file. Anybody wants to know more about the model itself, is recommended to read our paper (when it will be published).